// src/app/clientes/models/cliente-post.ts
export class ClientePost {
    nome: string;
    cpf: string;
    dataNascimento: string;
    telefone1: string;
    telefone2: string;
    email: string;
    observacao: string;
    logradouro: string;
    numero: string;
    complemento: string;
    condominio: string;
    bairro: string;
    municipio: string;
    estado: string;
    cep: string;
}